webhook/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
cachetools==5.5.0
camelot-py==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.14
dicttoxml==1.7.16
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.5
Flask==3.1.0
git-filter-repo==2.45.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-storage==2.10.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.1
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
priority==2.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymssql==2.3.1
PyMuPDF==1.24.13
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.41.2
tabula-py==2.10.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
zipp==3.21.0
